[황학선] #Fix Front : 실시간 트래킹 페이지 디버그 프린터 로거로 전환

See merge request s12-final/S12P31C104!204

+++ front/lib/models/hiking_route.dart
@@ -8,7 +8,7 @@ class HikingRoute {
   final num id;
   final num mountainId;
   final String name;
-  final double distance;
+  final int distance;
   final int estimatedTime; // 분 단위
   final String difficulty;
   final String description;
@@ -65,10 +65,10 @@ class HikingRoute {
           json['mountainId'] ?? json['mountain_id'] ?? 0, // mountain_id도 확인
       name: json['pathName'] ?? '',
       distance: (json['distance'] is num)
-          ? json['distance'].toDouble()
+          ? json['distance'].toInt()
           : (json['pathLength'] is num)
-              ? json['pathLength'].toDouble()
-              : 0.0,
+              ? json['pathLength'].toInt()
+              : 0,
       estimatedTime: json['estimatedTime'] is int
           ? json['estimatedTime']
           : (json['pathTime'] != null)
